fix: move mixin to monitoring
